[chore] moved sample into the samples folder
