--- Samples/Use Cache.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cfd64dc087600e447804b0527987b608
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
