--- Samples/Use Cache/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e5a7300f995672848bb4eccdc124d83c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
