--- Samples/Use Cache/Prefabs/Asset Page Controller.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c2d7e3ef517669b4a99f3aefb7f285bf
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
