--- Samples/Use Cache/Prefabs/Buttons.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f8dcb2fbdd8cc7b458b5038a0fcf81e1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
