--- Samples/Use Cache/Prefabs/Buttons/Asset Button.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 294089c67a283c0489aed948a7cb426b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
