--- Samples/Use Cache/Prefabs/Buttons/Category Button.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c11f9f095e187124487d29865bdfb62e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
