--- Samples/Use Cache/Prefabs/Category Controller.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0d255efc6c4bf2248a60ed78fddc645e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
