--- Samples/Use Cache/Prefabs/Paginator.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b63e03c00510f714eae6de80731f8c0c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
