--- Samples/Use Cache/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1bb56cd98a9b0234ca1d83c7fae8cda4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
