--- Samples/Use Cache/Scripts/Buttons.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dd2266a72cde18540a432609fc905d6f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
