--- Samples/Use Cache/Scripts/Buttons/AssetButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc6e54a57871f6c42ae5401928d3aa91
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
