--- Samples/Use Cache/Scripts/Buttons/CategoryButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 907e802302664c8438741db9f33a0b12
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
