--- Samples/Use Cache/Scripts/Components.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff9d842d21cb70a4fbbdec4b5c6c044d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
