--- Samples/Use Cache/Scripts/Components/AssetPageController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c1b47bf8ba9732e4697042a0a987cf28
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
