--- Samples/Use Cache/Scripts/Components/CategoryController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df19c509f7d7f3643a85407bc20a1aae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
