--- Samples/Use Cache/Scripts/Components/Paginator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5369a360f442fc44ca78d11c3c097c61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
