--- Samples/Use Cache/Scripts/EventAggregator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 77b5a760072929e488ccb8ab64645c72
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
