--- Samples/Use Cache/Scripts/UseCache.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f3e74301e31320345abc69b7641b7618
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
