--- Samples/Use Cache/Use Cache.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ebdc944e248f7b94b83163144d18a9a7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
