--- Samples/Use Cache/Visuals.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cb7a9b4a5fabea94b980e61e182c3fa9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
