--- Samples/Use Cache/Visuals/Icons.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5e5695788db1dbf47857ec1ea1d26bb4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
