+++ Samples~/Use Cache/Use Cache.unity
@@ -1446,8 +1446,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: -120}
-  m_SizeDelta: {x: 1400, y: 706}
+  m_AnchoredPosition: {x: 0, y: -62.500008}
+  m_SizeDelta: {x: 1400, y: 821}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1517004369
 MonoBehaviour:
@@ -2143,6 +2143,7 @@ MonoBehaviour:
   assetsToggle: {fileID: 1125715150}
   charactersToggle: {fileID: 343242684}
   mainUI: {fileID: 1517004366}
+  topMenu: {fileID: 1764656888}
   characterPosition: {fileID: 655478734}
 --- !u!1 &2008010650
 GameObject:
